requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-docx
tavily-python
